A15: три переписи выучили ступень 21 — и BuiltGraphSource стал uint64, потому что перепись корней настояла
AssetRoots поймала настоящее: AssetHandle на компоненте — это ССЫЛКА, которую обязан
пометить обход выселения. BuiltGraphSource ссылкой не является (это отпечаток
последней сборки эвалюатора), и пометка держала бы граф живым после очистки слота.
Тип поменян на uint64 — ровно как у соседнего BuiltRigSource.

SceneTextKeySigilMigration больше не голова, и это нормальный конец жизни ступени.
Пиним теперь то, что двигаться НЕ должно: её собственный номер 20 навсегда.

PointerOwnership: 840 -> 843 (+2 Raw, +1 Shared), строка с доводом на каждую.

+++ Desert/Desert/Source/Engine/ECS/System/AnimationECSSystem.hpp
@@ -438,7 +438,7 @@ namespace Desert::ECS
             {
                 anim.Graph = asset->GetGraph();
                 // The evaluator was built from the OLD object; the revision compare below rebuilds it.
-                anim.BuiltGraphSource = wanted;
+                anim.BuiltGraphSource = static_cast<uint64_t>( wanted );
             }
             return asset->GetRevision();
         }
